fix about the repository and view model of inventory and login
