+++ app/src/main/java/com/example/equipoCinco/repository/InventoryRepository.kt
@@ -14,39 +14,38 @@ class InventoryRepository  @Inject constructor(
     private val inventoryDao: InventoryDao,
     private val apiService: ApiService,
     private val db: FirebaseFirestore,
-    private val db_auth: FirebaseAuth = FirebaseAuth.getInstance()
-){
-     suspend fun saveInventory(inventory:Inventory){
-         withContext(Dispatchers.IO){
-             db.collection("inventory").document(inventory.id.toString()).set(
-                 hashMapOf(
-                     "id" to inventory.id,
-                     "name" to inventory.name,
-                     "price" to inventory.price,
-                     "quantity" to inventory.quantity
-                 )
-             )
-         }
+) {
+    suspend fun saveInventory(inventory: Inventory) {
+        withContext(Dispatchers.IO) {
+            db.collection("inventory").document(inventory.id.toString()).set(
+                hashMapOf(
+                    "id" to inventory.id,
+                    "name" to inventory.name,
+                    "price" to inventory.price,
+                    "quantity" to inventory.quantity
+                )
+            )
+        }
 
-         /*withContext(Dispatchers.IO){
+        /*withContext(Dispatchers.IO){
              inventoryDao.saveInventory(inventory)
          }*/
-     }
+    }
 
-    suspend fun getListInventory():MutableList<Inventory>{
-        return withContext(Dispatchers.IO){
+    suspend fun getListInventory(): MutableList<Inventory> {
+        return withContext(Dispatchers.IO) {
             inventoryDao.getListInventory()
         }
     }
 
-    suspend fun deleteInventory(inventory: Inventory){
-        withContext(Dispatchers.IO){
+    suspend fun deleteInventory(inventory: Inventory) {
+        withContext(Dispatchers.IO) {
             inventoryDao.deleteInventory(inventory)
         }
     }
 
-    suspend fun updateRepositoy(inventory: Inventory){
-        withContext(Dispatchers.IO){
+    suspend fun updateRepositoy(inventory: Inventory) {
+        withContext(Dispatchers.IO) {
             inventoryDao.updateInventory(inventory)
         }
     }
@@ -63,22 +62,8 @@ class InventoryRepository  @Inject constructor(
             }
         }
     }
-
-     fun registerUser(email: String, pass:String, isRegisterComplete: (Boolean)->Unit){
-        if(email.isNotEmpty() && pass.isNotEmpty()){
-            db_auth.createUserWithEmailAndPassword(email,pass)
-                .addOnCompleteListener {
-                    if (it.isSuccessful) {
-                        isRegisterComplete(true)
-                    } else {
-                        isRegisterComplete(false)
-                    }
-                }
-        }else{
-            isRegisterComplete(false)
-        }
-    }
-
 }
+
+
