+++ app/src/main/java/com/example/equipoCinco/repository/LoginRepository.kt
@@ -1,0 +1,22 @@
+package com.example.equipoCinco.repository
+
+import com.google.firebase.auth.FirebaseAuth
+
+class LoginRepository {
+    private val firebaseAuth = FirebaseAuth.getInstance()
+
+    fun registerUser(email: String, pass:String, isRegisterComplete: (Boolean)->Unit){
+        if(email.isNotEmpty() && pass.isNotEmpty()){
+            firebaseAuth.createUserWithEmailAndPassword(email,pass)
+                .addOnCompleteListener {
+                    if (it.isSuccessful) {
+                        isRegisterComplete(true)
+                    } else {
+                        isRegisterComplete(false)
+                    }
+                }
+        }else{
+            isRegisterComplete(false)
+        }
+    }
+}
